requirement.txt
django
gunicorn
mysqlclient
pycrypto
djangorestframework
djangorestframework-jwt
pygments
gensim
django-material
django-viewflow